changed ints into float64

+++ pkg/statistics/average.go
@@ -1,10 +1,10 @@
 package statistics
 
-func Average(numSet []int) int {
-	var average int
+func Average(numSet []float64) float64 {
+	var average float64
 	for _, num := range numSet{
 		average += num
 	}
-	average = average/len(numSet)
+	average = average/float64(len(numSet))
 	return average
 }
